[cpu] Add test_subproc.py to CI
It was giving "not compiled with CUDA" errors because the test was querying for
the CUDA compute capability, but it was not actually using this information. It
looks like the first iteration of the test needed to know the compute
capability, but that changed somewhere along the way.
