+++ python/test/unit/conftest.py
@@ -1,6 +1,6 @@
-# content of conftest.py
-
 import pytest
+import os
+import tempfile
 
 
 def pytest_addoption(parser):
@@ -10,3 +10,12 @@ def pytest_addoption(parser):
 @pytest.fixture
 def device(request):
     return request.config.getoption("--device")
+
+
+@pytest.fixture
+def fresh_triton_cache():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        try:
+            os.environ["TRITON_CACHE_DIR"] = tmpdir
+        finally:
+            os.environ.pop("TRITON_CACHE_DIR", None)
