+++ python/test/unit/runtime/test_subproc.py
@@ -2,8 +2,6 @@ import multiprocessing
 import os
 import shutil
 
-import torch
-
 import triton
 import triton.language as tl
 from triton.compiler import ASTSource
@@ -19,7 +17,7 @@ def reset_tmp_dir():
         shutil.rmtree(tmpdir, ignore_errors=True)
 
 
-def compile_fn(attrs, capability):
+def compile_fn(attrs):
 
     @triton.jit
     def kernel_sub(a, b, o, N: tl.constexpr):
@@ -36,18 +34,16 @@ def compile_fn(attrs, capability):
 
 
 def test_compile_in_subproc() -> None:
-    major, minor = torch.cuda.get_device_capability(0)
-    cc = major * 10 + minor
     config = triton.compiler.AttrsDescriptor(tuple(range(4)), ())
 
     multiprocessing.set_start_method('fork')
-    proc = multiprocessing.Process(target=compile_fn, args=(config, cc))
+    proc = multiprocessing.Process(target=compile_fn, args=(config, ))
     proc.start()
     proc.join()
     assert proc.exitcode == 0
 
 
-def compile_fn_dot(attrs, capability):
+def compile_fn_dot(attrs):
 
     @triton.jit
     def kernel_dot(Z):
@@ -60,14 +56,11 @@ def compile_fn_dot(attrs, capability):
     triton.compile(src=src, target=target)
 
 
-def test_compile_in_forked_subproc() -> None:
-    reset_tmp_dir()
-    major, minor = torch.cuda.get_device_capability(0)
-    capability = major * 10 + minor
+def test_compile_in_forked_subproc(fresh_triton_cache) -> None:
     config = triton.compiler.AttrsDescriptor(tuple(range(1)), ())
 
     assert multiprocessing.get_start_method() == 'fork'
-    proc = multiprocessing.Process(target=compile_fn_dot, args=(config, capability))
+    proc = multiprocessing.Process(target=compile_fn_dot, args=(config, ))
     proc.start()
     proc.join()
     assert proc.exitcode == 0
